refactor: ♻️ move task components to domain directory
